backend/pyproject.toml
[project]
name = "book-voice-backend"
version = "0.1.0"
description = "FastAPI backend for book page OCR + TTS."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pillow>=11.0.0,<12.0.0",
    "numpy>=2.1.0,<3.0.0",
    "paddleocr>=3.1.0,<4.0.0",
    "paddlepaddle>=3.0.0,<4.0.0",
    "kokoro==0.9.4",
    "torch==2.11.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.4.2,<9.0.0",
]

[tool.uv.sources]
paddlepaddle = [{ index = "paddle-cpu" }]
torch = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "paddle-cpu"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
